LayoutTests/inspector/debugger/resource-script-mapping-expected.txt

Tests ResourceScriptMapping class.


Running: testScriptWithPendingResource
Adding script for pending request.
UISourceCode: foo.js
UISourceCode is editable: false
UISourceCode is content script: true
Mime type: text/javascript
UISourceCode content: <content script source>
Adding uiSourceCode for finished resource.
UISourceCode: foo.js
UISourceCode is editable: true
UISourceCode is content script: false
Mime type: text/javascript
UISourceCode content: <content script resource content>

Running: testScriptWithFinishedResource
Adding uiSourceCode for finished resource.
UISourceCode: foo.js
UISourceCode is editable: true
UISourceCode is content script: false
Mime type: text/javascript
UISourceCode content: <script resource content>
Adding script for finished request.
UISourceCode: foo.js
UISourceCode is editable: true
UISourceCode is content script: false
Mime type: text/javascript
UISourceCode content: <script resource content>

Running: testHTMLWithPendingResource
Adding first script for pending request.
UISourceCode: index.html
UISourceCode is editable: false
UISourceCode is content script: false
Mime type: text/html
UISourceCode content:   <script><script source 1></script>
Adding second script for pending request.
UISourceCode: index.html
UISourceCode is editable: false
UISourceCode is content script: false
Mime type: text/html
UISourceCode content:                                      <script><script source 2></script>
Adding uiSourceCode for finished resource.
UISourceCode: index.html
UISourceCode is editable: false
UISourceCode is content script: false
Mime type: text/html
UISourceCode content: <resource content>

Running: testHTMLWithFinishedResource
Adding uiSourceCode for finished resource.
UISourceCode: index.html
UISourceCode is editable: false
UISourceCode is content script: false
Mime type: text/html
UISourceCode content: <resource content>
Adding first script for finished request.
UISourceCode: index.html
UISourceCode is editable: false
UISourceCode is content script: false
Mime type: text/html
UISourceCode content: <resource content>
Adding second script for finished request.
UISourceCode: index.html
UISourceCode is editable: false
UISourceCode is content script: false
Mime type: text/html
UISourceCode content: <resource content>
